Move code out of recompute folder and into root
